requirements.txt
Pyrebase4==4.6.0